% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtpsImportaDipr.R
\name{mtpsImportaDipr}
\alias{mtpsImportaDipr}
\title{Importacao de dados contidos no Demonstrativo de Informacoes Previdenciarias e Repasses (DIPR)}
\usage{
mtpsImportaDipr(arqNome, arqTipo, opcao)
}
\arguments{
\item{arqNome}{uma string contendo o nome do arquivo pdf correspondente ao "DIPR".}

\item{arqTipo}{uma string contendo o tipo do arquivo de origem (pdf ou xml)}

\item{opcao}{lista com as opcoes desejadas. Ver detalhes e exemplos abaixo.}
}
\value{
uma lista de objetos do tipo data.frame contendo as tabelas desejadas.
}
\description{
Esta funcao extrai os dados contidos no DIPR dos Regimes Proprios de Previdencia Social disponibilizado,
em formato pdf, no site do Min. do Trabalho e Previdencia Social
}
\details{
A funcao extrai tabelas contidas no DIPR de acordo com o estabelecido no parametro \code{opcao}:
OBS.: Por enquanto a função só extrai de arquivos do tipo "pdf.
\itemize{
  \item \code{opcao = 1} extrai tabela contendo as BASES DE CALCULO DAS CONTRIBUICOES DEVIDAS, RELATIVAS AS FOLHAS DO ENTE.
  \item \code{opcao = 2} extrai tabela contendo as CONTRIBUICOES REPASSADAS.
  \item \code{opcao = 3} extrai tabela contendo as DEDUCOES.
  \item \code{opcao = 4} extrai tabela contendo os APORTES E TRANSFERENCIAS DE RECURSOS.
  \item \code{opcao = 5} extrai tabela contendo os PARCELAMENTOS.
  \item \code{opcao = 6} extrai tabela contendo as BASES DE CALCULO DAS CONTRIBUICOES DEVIDAS, RELATIVAS A FOLHA DA UNIDADE GESTORA
  \item \code{opcao = 7} extrai tabela contendo as CONTRIBUICOES ARRECADADAS PELA UNIDADE GESTORA
  \item \code{opcao = 8} extrai tabela contendo a REMUNERACAO BRUTA (somatorio das folhas do ENTE e da UNIDADE GESTORA)
  \item \code{opcao = 9} extrai tabela contendo o Nº DE BENEFICIARIOS (somatorio das folhas do ENTE e da UNIDADE GESTORA)
}
}
\examples{
\dontrun{
tabList <- mtpsImportaDpin("arqNome = DIPR.pdf", arqTipo = "pdf", opcao = 1)

tabList <- mtpsImportaDpin("arqNome = DIPR.pdf", arqTipo = "pdf", opcao = c(1,3,4)
}
}
\author{
Bruno M. S. S Melo
}
\seealso{
\code{tabulizer::extract_tables}
}

