% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtpsImportaDiprTab1.R
\name{mtpsImportaDiprTab1}
\alias{mtpsImportaDiprTab1}
\title{Importacao de dados contidos no Demonstrativo de Informacoes Previdenciarias e Repasses (DIPR)}
\usage{
mtpsImportaDiprTab1(lsExtractedDipr)
}
\arguments{
\item{lsExtractedDipr}{saida da funcao \code{tabulizer::extract_tables} aplicada num arquivo pdf contendo a DIPR.}
}
\value{
um objeto do tipo data.frame contendo a tabela desejada.
}
\description{
Esta funcao extrai os dados contidos no DIPR dos Regimes Proprios de Previdencia Social disponibilizado,
em formato pdf, no site do Min. do Trabalho e Previdencia Social
}
\details{
extrai tabela contendo as BASES DE CALCULO DAS CONTRIBUICOES DEVIDAS, RELATIVAS AS FOLHAS DO ENTE.
}
\author{
Bruno M. S. S Melo
}

