% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtpsImportaDair.R
\name{mtpsImportaDair}
\alias{mtpsImportaDair}
\title{Importacao de dados contidos no DAIR disponiblizado no site do Min. do Trabalho e Previdencia Social.}
\usage{
mtpsImportaDair(caminho, tipo)
}
\arguments{
\item{caminho}{uma string contendo o nome do arquivo correspondente ao "DAIR". Caso o parametro "tipo"
seja "dir.pdf" ou "dir.xml", devera ser passada uma lista de strings contendo o caminho para um
ou mais diretorios.}

\item{tipo}{uma string contendo o tipo do arquivo de origem (pdf ou xml)}
}
\value{
um objeto do tipo data.frame contendo a relacao das aplicacoes financeiras do RPPS.
}
\description{
Esta funcao extrai os dados contidos no Demonstrativo de Aplicacoes e Investimentos dos Recursos (DAIR)
dos Regimes Proprios de Previdencia Social disponibilizado, em formato pdf, no site  do Min. do Trabalho e
Previdencia Social.
}
\details{
O parametro \code{tipo} indica o tipo da origem:
\itemize{
  \item \code{tipo = "pdf"} o \code{caminho} aponta para um arquivo do tipo pdf
  \item \code{tipo = "dir.pdf"} o \code{caminho} aponta para um diretorio contendo arquivos do tipo pdf
  \item \code{tipo = "xml"} o \code{caminho} aponta para um arquivo do tipo xml
  \item \code{tipo = "dir.xml"} o \code{caminho} aponta para um diretorio contendo arquivos do tipo xml
}
Obs. 1: Na presente versao apenas as aplicacoes financeiras realizadas em fundos de investimento sao extraidas.

Obs. 2: Por enquanto apenas as origens dos tipos "pdf" e "dir.pdf" estao implementadas.

Problema conhecido: Numa mesma sessao do R deve-se evitar a importacao de arquivos pdf de mesmo nome,
ainda que estejam em diretorios distintos, dado que apenas os dados do primeiro arquivo "homonino" serao
importados.
}
\examples{
\dontrun{
dfDair <- mtpsImportaDair(caminho = "DAIR.pdf", tipo = "pdf")
}
}
\author{
Bruno M. S. S. Melo
}
\seealso{
\code{tabulizer::extract_tables}
}

