% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtpsImportaDair.R
\name{mtpsImportaDair}
\alias{mtpsImportaDair}
\title{Importacao de dados contidos no DAIR disponiblizado no site do Min. do Trabalho e Previdencia Social.}
\usage{
mtpsImportaDair(arqNome)
}
\arguments{
\item{arqNome}{uma string contendo o nome do arquivo correspondente ao "DAIR".}

\item{arqTipo}{uma string contendo o tipo do arquivo de origem (pdf ou xml)}
}
\value{
um objeto do tipo data.frame contendo a relacao das aplicacoes financeiras do RPPS.
}
\description{
Esta funcao extrai os dados contidos no Demonstrativo de Aplicacoes e Investimentos dos Recursos (DAIR)
dos Regimes Proprios de Previdencia Social disponibilizado, em formato pdf, no site  do Min. do Trabalho e
Previdencia Social.
}
\details{
Na presente versao apenas as aplicacoes financeiras realizadas em fundos de investimento sao extraidas.
OBS.: Por enquanto a função só extrai de arquivos do tipo "pdf.
}
\examples{
\dontrun{
dfDair <- mtpsImportaDair("DAIR.pdf")
}
}
\author{
Bruno M. S. S. Melo
}
\seealso{
\code{tabulizer::extract_tables}
}

