% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtpsImportaDpin.R
\name{mtpsImportaDpin}
\alias{mtpsImportaDpin}
\title{Importacao de dados contidos no Demonstrativo da Politica de Investimentos (DPIN)}
\usage{
mtpsImportaDpin(arqNome, arqTipo)
}
\arguments{
\item{arqNome}{uma string contendo o nome do arquivo correspondente ao "DPIN".}

\item{arqTipo}{uma string contendo o tipo do arquivo de origem (pdf ou xml)}
}
\value{
um objeto do tipo data.frame contendo o resumo da politica de investimentos do RPPS.
}
\description{
Esta funcao extrai os dados contidos no Demonstrativo da Politica de Investimentos (DPIN)
dos Regimes Proprios de Previdencia Social disponibilizado, em formato pdf, nos sites de alguns
Institutos de Previdencia de entes federativos que possuem RPPS.
}
\details{
A funcao extrai a tabela contendo o resumo da alocacao de recursos definida no "DPIN".
OBS.: Por enquanto a função só extrai de arquivos do tipo "pdf.
}
\examples{
\dontrun{
dfDpin <- mtpsImportaDpin("DPIN.pdf")
}
}
\author{
Bruno M. S. S Melo
}
\seealso{
\code{tabulizer::extract_tables}
}

