% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtpsImportaDpin.R
\name{mtpsImportaDpin}
\alias{mtpsImportaDpin}
\title{Importacao de dados contidos no Demonstrativo da Politica de Investimentos (DPIN)}
\usage{
mtpsImportaDpin(arqDpin)
}
\arguments{
\item{arqDpin}{- uma string contendo o nome do arquivo pdf correspondente ao "DPIN".}
}
\value{
um objeto do tipo data.frame contendo a politica de alocacao das aplicacoes financeiras do RPPS.
}
\description{
Esta funcao extrai os dados contidos no Demonstrativo da Politica de Investimentos (DPIN)
dos Regimes Proprios de Previdencia Social disponibilizado, em formato pdf, nos sites de alguns
Institutos de Previdencia de entes federativos que possuem RPPS.
}
\details{
Ainda em desenvolvimento.
}
\author{
Bruno M. S. S Melo
}
\seealso{
\code{extract_tables}
}

